backend: Split Condition of Sale tests to separate files
