+++ backend/hitas/tests/apis/condition_of_sale/test_api_condition_of_sale_create.py
@@ -1,6 +1,5 @@
 import datetime
-import uuid
-from typing import Any, NamedTuple, TypedDict
+from typing import Any
 
 import pytest
 from dateutil.relativedelta import relativedelta
@@ -8,7 +7,7 @@ from django.urls import reverse
 from django.utils import timezone
 from rest_framework import status
 
-from hitas.models.condition_of_sale import ConditionOfSale, GracePeriod
+from hitas.models.condition_of_sale import ConditionOfSale
 from hitas.models.housing_company import HitasType, RegulationStatus
 from hitas.models.owner import Owner
 from hitas.models.ownership import Ownership
@@ -18,401 +17,6 @@ from hitas.tests.factories import (
     OwnerFactory,
     OwnershipFactory,
 )
-
-# List tests
-
-
-@pytest.mark.django_db
-def test__api__condition_of_sale__list__empty(api_client: HitasAPIClient):
-    OwnerFactory.create()
-
-    url = reverse("hitas:conditions-of-sale-list")
-    response = api_client.get(url)
-    assert response.status_code == status.HTTP_200_OK, response.json()
-    assert response.json()["contents"] == []
-    assert response.json()["page"] == {
-        "size": 0,
-        "current_page": 1,
-        "total_items": 0,
-        "total_pages": 1,
-        "links": {
-            "next": None,
-            "previous": None,
-        },
-    }
-
-
-@pytest.mark.django_db
-def test__api__condition_of_sale__list__single(api_client: HitasAPIClient):
-    cos: ConditionOfSale = ConditionOfSaleFactory.create()
-
-    url = reverse("hitas:conditions-of-sale-list")
-    response = api_client.get(url)
-    assert response.status_code == status.HTTP_200_OK, response.json()
-    assert response.json()["contents"] == [
-        {
-            "id": cos.uuid.hex,
-            "grace_period": str(cos.grace_period.value),
-            "fulfilled": cos.fulfilled,
-            "new_ownership": {
-                "percentage": float(cos.new_ownership.percentage),
-                "apartment": {
-                    "id": cos.new_ownership.apartment.uuid.hex,
-                    "address": {
-                        "street_address": cos.new_ownership.apartment.street_address,
-                        "apartment_number": cos.new_ownership.apartment.apartment_number,
-                        "floor": cos.new_ownership.apartment.floor,
-                        "stair": cos.new_ownership.apartment.stair,
-                        "city": "Helsinki",
-                        "postal_code": cos.new_ownership.apartment.postal_code.value,
-                    },
-                    "housing_company": {
-                        "id": cos.new_ownership.apartment.housing_company.uuid.hex,
-                        "display_name": cos.new_ownership.apartment.housing_company.display_name,
-                    },
-                },
-                "owner": {
-                    "id": cos.new_ownership.owner.uuid.hex,
-                    "name": cos.new_ownership.owner.name,
-                    "identifier": cos.new_ownership.owner.identifier,
-                    "email": cos.new_ownership.owner.email,
-                    "non_disclosure": cos.new_ownership.owner.non_disclosure,
-                },
-            },
-            "old_ownership": {
-                "percentage": float(cos.old_ownership.percentage),
-                "apartment": {
-                    "id": cos.old_ownership.apartment.uuid.hex,
-                    "address": {
-                        "street_address": cos.old_ownership.apartment.street_address,
-                        "apartment_number": cos.old_ownership.apartment.apartment_number,
-                        "floor": cos.old_ownership.apartment.floor,
-                        "stair": cos.old_ownership.apartment.stair,
-                        "city": "Helsinki",
-                        "postal_code": cos.old_ownership.apartment.postal_code.value,
-                    },
-                    "housing_company": {
-                        "id": cos.old_ownership.apartment.housing_company.uuid.hex,
-                        "display_name": cos.old_ownership.apartment.housing_company.display_name,
-                    },
-                },
-                "owner": {
-                    "id": cos.old_ownership.owner.uuid.hex,
-                    "name": cos.old_ownership.owner.name,
-                    "identifier": cos.old_ownership.owner.identifier,
-                    "email": cos.old_ownership.owner.email,
-                    "non_disclosure": cos.old_ownership.owner.non_disclosure,
-                },
-            },
-        }
-    ]
-    assert response.json()["page"] == {
-        "size": 1,
-        "current_page": 1,
-        "total_items": 1,
-        "total_pages": 1,
-        "links": {
-            "next": None,
-            "previous": None,
-        },
-    }
-
-
-@pytest.mark.django_db
-def test__api__condition_of_sale__list__multiple(api_client: HitasAPIClient):
-    cos_1: ConditionOfSale = ConditionOfSaleFactory.create()
-    cos_2: ConditionOfSale = ConditionOfSaleFactory.create()
-
-    url = reverse("hitas:conditions-of-sale-list")
-    response = api_client.get(url)
-    assert response.status_code == status.HTTP_200_OK, response.json()
-    assert response.json()["contents"] == [
-        {
-            "id": cos_1.uuid.hex,
-            "grace_period": str(cos_1.grace_period.value),
-            "fulfilled": cos_1.fulfilled,
-            "new_ownership": {
-                "percentage": float(cos_1.new_ownership.percentage),
-                "apartment": {
-                    "id": cos_1.new_ownership.apartment.uuid.hex,
-                    "address": {
-                        "street_address": cos_1.new_ownership.apartment.street_address,
-                        "apartment_number": cos_1.new_ownership.apartment.apartment_number,
-                        "floor": cos_1.new_ownership.apartment.floor,
-                        "stair": cos_1.new_ownership.apartment.stair,
-                        "city": "Helsinki",
-                        "postal_code": cos_1.new_ownership.apartment.postal_code.value,
-                    },
-                    "housing_company": {
-                        "id": cos_1.new_ownership.apartment.housing_company.uuid.hex,
-                        "display_name": cos_1.new_ownership.apartment.housing_company.display_name,
-                    },
-                },
-                "owner": {
-                    "id": cos_1.new_ownership.owner.uuid.hex,
-                    "name": cos_1.new_ownership.owner.name,
-                    "identifier": cos_1.new_ownership.owner.identifier,
-                    "email": cos_1.new_ownership.owner.email,
-                    "non_disclosure": cos_1.new_ownership.owner.non_disclosure,
-                },
-            },
-            "old_ownership": {
-                "percentage": float(cos_1.old_ownership.percentage),
-                "apartment": {
-                    "id": cos_1.old_ownership.apartment.uuid.hex,
-                    "address": {
-                        "street_address": cos_1.old_ownership.apartment.street_address,
-                        "apartment_number": cos_1.old_ownership.apartment.apartment_number,
-                        "floor": cos_1.old_ownership.apartment.floor,
-                        "stair": cos_1.old_ownership.apartment.stair,
-                        "city": "Helsinki",
-                        "postal_code": cos_1.old_ownership.apartment.postal_code.value,
-                    },
-                    "housing_company": {
-                        "id": cos_1.old_ownership.apartment.housing_company.uuid.hex,
-                        "display_name": cos_1.old_ownership.apartment.housing_company.display_name,
-                    },
-                },
-                "owner": {
-                    "id": cos_1.old_ownership.owner.uuid.hex,
-                    "name": cos_1.old_ownership.owner.name,
-                    "identifier": cos_1.old_ownership.owner.identifier,
-                    "email": cos_1.old_ownership.owner.email,
-                    "non_disclosure": cos_1.old_ownership.owner.non_disclosure,
-                },
-            },
-        },
-        {
-            "id": cos_2.uuid.hex,
-            "grace_period": str(cos_2.grace_period.value),
-            "fulfilled": cos_2.fulfilled,
-            "new_ownership": {
-                "percentage": float(cos_2.new_ownership.percentage),
-                "apartment": {
-                    "id": cos_2.new_ownership.apartment.uuid.hex,
-                    "address": {
-                        "street_address": cos_2.new_ownership.apartment.street_address,
-                        "apartment_number": cos_2.new_ownership.apartment.apartment_number,
-                        "floor": cos_2.new_ownership.apartment.floor,
-                        "stair": cos_2.new_ownership.apartment.stair,
-                        "city": "Helsinki",
-                        "postal_code": cos_2.new_ownership.apartment.postal_code.value,
-                    },
-                    "housing_company": {
-                        "id": cos_2.new_ownership.apartment.housing_company.uuid.hex,
-                        "display_name": cos_2.new_ownership.apartment.housing_company.display_name,
-                    },
-                },
-                "owner": {
-                    "id": cos_2.new_ownership.owner.uuid.hex,
-                    "name": cos_2.new_ownership.owner.name,
-                    "identifier": cos_2.new_ownership.owner.identifier,
-                    "email": cos_2.new_ownership.owner.email,
-                    "non_disclosure": cos_2.new_ownership.owner.non_disclosure,
-                },
-            },
-            "old_ownership": {
-                "percentage": float(cos_2.old_ownership.percentage),
-                "apartment": {
-                    "id": cos_2.old_ownership.apartment.uuid.hex,
-                    "address": {
-                        "street_address": cos_2.old_ownership.apartment.street_address,
-                        "apartment_number": cos_2.old_ownership.apartment.apartment_number,
-                        "floor": cos_2.old_ownership.apartment.floor,
-                        "stair": cos_2.old_ownership.apartment.stair,
-                        "city": "Helsinki",
-                        "postal_code": cos_2.old_ownership.apartment.postal_code.value,
-                    },
-                    "housing_company": {
-                        "id": cos_2.old_ownership.apartment.housing_company.uuid.hex,
-                        "display_name": cos_2.old_ownership.apartment.housing_company.display_name,
-                    },
-                },
-                "owner": {
-                    "id": cos_2.old_ownership.owner.uuid.hex,
-                    "name": cos_2.old_ownership.owner.name,
-                    "identifier": cos_2.old_ownership.owner.identifier,
-                    "email": cos_2.old_ownership.owner.email,
-                    "non_disclosure": cos_2.old_ownership.owner.non_disclosure,
-                },
-            },
-        },
-    ]
-    assert response.json()["page"] == {
-        "size": 2,
-        "current_page": 1,
-        "total_items": 2,
-        "total_pages": 1,
-        "links": {
-            "next": None,
-            "previous": None,
-        },
-    }
-
-
-@pytest.mark.django_db
-def test__api__condition_of_sale__list__fulfilled_under_x_months(api_client: HitasAPIClient, freezer, settings):
-    freezer.move_to("2023-01-01 00:00:00+00:00")
-    old_time = timezone.now()
-    old_time_str = old_time.replace(tzinfo=None).isoformat(timespec="seconds") + "Z"
-
-    ConditionOfSaleFactory.create()
-    condition_of_sale_2: ConditionOfSale = ConditionOfSaleFactory.create()
-    condition_of_sale_2.delete()
-
-    freezer.move_to(old_time + settings.SHOW_FULFILLED_CONDITIONS_OF_SALE_FOR_MONTHS)
-
-    url = reverse("hitas:conditions-of-sale-list")
-    response = api_client.get(url)
-    assert response.status_code == status.HTTP_200_OK, response.json()
-    assert len(response.json()["contents"]) == 2
-    assert response.json()["contents"][1]["fulfilled"] == old_time_str
-
-
-@pytest.mark.django_db
-def test__api__condition_of_sale__list__fulfilled_over_x_months(api_client: HitasAPIClient, freezer, settings):
-    freezer.move_to("2023-01-01 00:00:00+00:00")
-    old_time = timezone.now()
-
-    condition_of_sale_1: ConditionOfSale = ConditionOfSaleFactory.create()
-    condition_of_sale_2: ConditionOfSale = ConditionOfSaleFactory.create()
-    condition_of_sale_2.delete()
-
-    freezer.move_to(old_time + settings.SHOW_FULFILLED_CONDITIONS_OF_SALE_FOR_MONTHS + relativedelta(seconds=1))
-
-    url = reverse("hitas:conditions-of-sale-list")
-    response = api_client.get(url)
-    assert response.status_code == status.HTTP_200_OK, response.json()
-    assert len(response.json()["contents"]) == 1
-
-    alive_conditions_of_sale: list[ConditionOfSale] = list(ConditionOfSale.objects.all())
-    assert len(alive_conditions_of_sale) == 1
-    assert alive_conditions_of_sale[0].uuid.hex == condition_of_sale_1.uuid.hex
-    assert alive_conditions_of_sale[0].fulfilled is None
-
-    deleted_conditions_of_sale: list[ConditionOfSale] = list(ConditionOfSale.deleted_objects.all())
-    assert len(deleted_conditions_of_sale) == 1
-    assert deleted_conditions_of_sale[0].uuid.hex == condition_of_sale_2.uuid.hex
-    assert deleted_conditions_of_sale[0].fulfilled == old_time
-
-
-# Retrieve tests
-
-
-@pytest.mark.django_db
-def test__api__condition_of_sale__retrieve(api_client: HitasAPIClient):
-    cos: ConditionOfSale = ConditionOfSaleFactory.create()
-
-    url = reverse(
-        "hitas:conditions-of-sale-detail",
-        kwargs={
-            "uuid": cos.uuid.hex,
-        },
-    )
-    response = api_client.get(url)
-    assert response.status_code == status.HTTP_200_OK, response.json()
-    assert response.json() == {
-        "id": cos.uuid.hex,
-        "grace_period": str(cos.grace_period.value),
-        "fulfilled": cos.fulfilled,
-        "new_ownership": {
-            "percentage": float(cos.new_ownership.percentage),
-            "apartment": {
-                "id": cos.new_ownership.apartment.uuid.hex,
-                "address": {
-                    "street_address": cos.new_ownership.apartment.street_address,
-                    "apartment_number": cos.new_ownership.apartment.apartment_number,
-                    "floor": cos.new_ownership.apartment.floor,
-                    "stair": cos.new_ownership.apartment.stair,
-                    "city": "Helsinki",
-                    "postal_code": cos.new_ownership.apartment.postal_code.value,
-                },
-                "housing_company": {
-                    "id": cos.new_ownership.apartment.housing_company.uuid.hex,
-                    "display_name": cos.new_ownership.apartment.housing_company.display_name,
-                },
-            },
-            "owner": {
-                "id": cos.new_ownership.owner.uuid.hex,
-                "name": cos.new_ownership.owner.name,
-                "identifier": cos.new_ownership.owner.identifier,
-                "email": cos.new_ownership.owner.email,
-                "non_disclosure": cos.new_ownership.owner.non_disclosure,
-            },
-        },
-        "old_ownership": {
-            "percentage": float(cos.old_ownership.percentage),
-            "apartment": {
-                "id": cos.old_ownership.apartment.uuid.hex,
-                "address": {
-                    "street_address": cos.old_ownership.apartment.street_address,
-                    "apartment_number": cos.old_ownership.apartment.apartment_number,
-                    "floor": cos.old_ownership.apartment.floor,
-                    "stair": cos.old_ownership.apartment.stair,
-                    "city": "Helsinki",
-                    "postal_code": cos.old_ownership.apartment.postal_code.value,
-                },
-                "housing_company": {
-                    "id": cos.old_ownership.apartment.housing_company.uuid.hex,
-                    "display_name": cos.old_ownership.apartment.housing_company.display_name,
-                },
-            },
-            "owner": {
-                "id": cos.old_ownership.owner.uuid.hex,
-                "name": cos.old_ownership.owner.name,
-                "identifier": cos.old_ownership.owner.identifier,
-                "email": cos.old_ownership.owner.email,
-                "non_disclosure": cos.old_ownership.owner.non_disclosure,
-            },
-        },
-    }
-
-
-@pytest.mark.django_db
-def test__api__condition_of_sale__retrieve__fulfilled_over_x_months(api_client: HitasAPIClient, freezer, settings):
-    freezer.move_to("2023-01-01 00:00:00+00:00")
-    old_time = timezone.now()
-
-    condition_of_sale: ConditionOfSale = ConditionOfSaleFactory.create()
-    condition_of_sale.delete()
-
-    freezer.move_to(old_time + settings.SHOW_FULFILLED_CONDITIONS_OF_SALE_FOR_MONTHS + relativedelta(seconds=1))
-
-    url = reverse(
-        "hitas:conditions-of-sale-detail",
-        kwargs={
-            "uuid": condition_of_sale.uuid.hex,
-        },
-    )
-    response = api_client.get(url)
-    assert response.status_code == status.HTTP_404_NOT_FOUND, response.json()
-    assert response.json() == {
-        "error": "condition_of_sale_not_found",
-        "message": "Condition of Sale not found",
-        "reason": "Not Found",
-        "status": 404,
-    }
-
-
-@pytest.mark.django_db
-def test__api__condition_of_sale__retrieve__invalid_id(api_client: HitasAPIClient):
-    url = reverse(
-        "hitas:conditions-of-sale-detail",
-        kwargs={
-            "uuid": uuid.uuid4().hex,
-        },
-    )
-    response = api_client.get(url)
-    assert response.status_code == status.HTTP_404_NOT_FOUND, response.json()
-    assert response.json() == {
-        "error": "condition_of_sale_not_found",
-        "message": "Condition of Sale not found",
-        "reason": "Not Found",
-        "status": 404,
-    }
-
-
-# Create tests
 
 
 @pytest.mark.django_db
@@ -1294,173 +898,3 @@ def test__api__condition_of_sale__create__not_to_half_hitas(api_client: HitasAPI
     assert len(response.json().get("conditions_of_sale", [])) == 0, response.json()
     conditions_of_sale: list[ConditionOfSale] = list(ConditionOfSale.objects.all())
     assert len(conditions_of_sale) == 0
-
-
-# Update tests
-
-
-class ConditionOfSaleUpdateData(TypedDict):
-    grace_period: str
-
-
-class ConditionOfSaleUpdateArgs(NamedTuple):
-    data: ConditionOfSaleUpdateData
-
-
-@pytest.mark.parametrize(
-    **parametrize_helper(
-        {
-            "Change grace period to three months": ConditionOfSaleUpdateArgs(
-                data=ConditionOfSaleUpdateData(
-                    grace_period=GracePeriod.THREE_MONTHS.value,
-                ),
-            ),
-            "Change grace period to six months": ConditionOfSaleUpdateArgs(
-                data=ConditionOfSaleUpdateData(
-                    grace_period=GracePeriod.SIX_MONTHS.value,
-                ),
-            ),
-        },
-    ),
-)
-@pytest.mark.django_db
-def test__api__condition_of_sale__update(api_client: HitasAPIClient, data: ConditionOfSaleUpdateData):
-    condition_of_sale: ConditionOfSale = ConditionOfSaleFactory.create(grace_period=GracePeriod.NOT_GIVEN)
-
-    url = reverse(
-        "hitas:conditions-of-sale-detail",
-        kwargs={
-            "uuid": condition_of_sale.uuid.hex,
-        },
-    )
-    response = api_client.put(url, data=data, format="json")
-
-    assert response.status_code == status.HTTP_200_OK, response.json()
-    assert response.json()["grace_period"] == str(data["grace_period"])
-    condition_of_sale.refresh_from_db()
-    assert condition_of_sale.grace_period.value == str(data["grace_period"])
-
-
-@pytest.mark.parametrize(
-    **parametrize_helper(
-        {
-            "Grace period cannot be null": InvalidInput(
-                invalid_data={"grace_period": None},
-                fields=[
-                    {
-                        "field": "grace_period",
-                        "message": "This field is mandatory and cannot be null.",
-                    },
-                ],
-            ),
-            "Not a valid value for grace period": InvalidInput(
-                invalid_data={"grace_period": "foo"},
-                fields=[
-                    {
-                        "field": "grace_period",
-                        "message": '"foo" is not a valid choice.',
-                    },
-                ],
-            ),
-        },
-    ),
-)
-@pytest.mark.django_db
-def test__api__condition_of_sale__update__invalid(api_client: HitasAPIClient, invalid_data: dict, fields: list):
-    condition_of_sale: ConditionOfSale = ConditionOfSaleFactory.create(grace_period=GracePeriod.NOT_GIVEN)
-
-    url = reverse(
-        "hitas:conditions-of-sale-detail",
-        kwargs={
-            "uuid": condition_of_sale.uuid.hex,
-        },
-    )
-    response = api_client.put(url, data=invalid_data, format="json", openapi_validate_request=False)
-    assert response.status_code == status.HTTP_400_BAD_REQUEST, response.json()
-    assert response.json() == {
-        "error": "bad_request",
-        "fields": fields,
-        "message": "Bad request",
-        "reason": "Bad Request",
-        "status": 400,
-    }
-
-
-# Delete tests
-
-
-@pytest.mark.django_db
-def test__api__condition_of_sale__delete__different_owners(api_client: HitasAPIClient, freezer, settings):
-    freezer.move_to("2023-01-01 00:00:00+00:00")
-
-    owner_1: Owner = OwnerFactory.create()
-    owner_2: Owner = OwnerFactory.create()
-
-    condition_of_sale: ConditionOfSale = ConditionOfSaleFactory.create(
-        new_ownership__owner=owner_1,
-        old_ownership__owner=owner_2,
-    )
-
-    url = reverse(
-        "hitas:conditions-of-sale-detail",
-        kwargs={
-            "uuid": condition_of_sale.uuid.hex,
-        },
-    )
-    response_1 = api_client.delete(url)
-    # Deletion is allowed since the old and new owners are not the same (created through a household)
-    assert response_1.status_code == status.HTTP_204_NO_CONTENT, response_1.json()
-
-    # The condition of sale is hard deleted
-    response_2 = api_client.get(url)
-    assert response_2.status_code == status.HTTP_404_NOT_FOUND, response_2.json()
-
-
-@pytest.mark.django_db
-def test__api__condition_of_sale__delete__same_owners(api_client: HitasAPIClient, settings):
-    owner: Owner = OwnerFactory.create()
-
-    condition_of_sale: ConditionOfSale = ConditionOfSaleFactory.create(
-        new_ownership__owner=owner,
-        old_ownership__owner=owner,
-    )
-
-    url = reverse(
-        "hitas:conditions-of-sale-detail",
-        kwargs={
-            "uuid": condition_of_sale.uuid.hex,
-        },
-    )
-    response = api_client.delete(url)
-    # Deletion is not allowed since the old and new owners are the same (created through a sale)
-    assert response.status_code == status.HTTP_409_CONFLICT, response.json()
-
-    assert response.json() == {
-        "error": "invalid",
-        "message": "Cannot delete condition of sale between the same owner.",
-        "reason": "Conflict",
-        "status": 409,
-    }
-
-
-@pytest.mark.django_db
-def test__api__condition_of_sale__delete__cascade_from_ownership(api_client: HitasAPIClient):
-    condition_of_sale: ConditionOfSale = ConditionOfSaleFactory.create()
-
-    assert condition_of_sale.fulfilled is None
-
-    condition_of_sale.new_ownership.delete()
-    condition_of_sale.refresh_from_db()
-    assert condition_of_sale.fulfilled is not None
-
-    condition_of_sale.new_ownership.undelete()
-    condition_of_sale.refresh_from_db()
-    assert condition_of_sale.fulfilled is None
-
-    condition_of_sale.old_ownership.delete()
-    condition_of_sale.refresh_from_db()
-    assert condition_of_sale.fulfilled is not None
-
-    condition_of_sale.old_ownership.undelete()
-    condition_of_sale.refresh_from_db()
-    assert condition_of_sale.fulfilled is None
